cdk/requirements.txt
aws-cdk-lib==2.136.0
constructs>=10.0.0,<11.0.0
awscli-local==0.22.0
boto3>=1.34.0
dotenv